flower-recognition-backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.0
python-multipart==0.0.20
requests==2.32.3
python-dotenv==1.0.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.9